Move fmt integrtion header out of detail namespace
Also polishing the implementation to match the previous output.
