+++ libvast/vast/concept/printable/vast/view.hpp
@@ -108,12 +108,12 @@ struct record_view_printer : printer<record_view_printer> {
   }
 };
 
-// -- printer registry setup ---------------------------------------------------
+// // -- printer registry setup ---------------------------------------------------
 
-template <>
-struct printer_registry<view<data>> {
-  using type = data_view_printer;
-};
+// template <>
+// struct printer_registry<view<data>> {
+//   using type = data_view_printer;
+// };
 
 template <>
 struct printer_registry<view<pattern>> {
