+++ libvast/vast/data.hpp
@@ -13,7 +13,6 @@
 #include "vast/concept/hashable/uhash.hpp"
 #include "vast/concept/hashable/xxhash.hpp"
 #include "vast/defaults.hpp"
-#include "vast/detail/fmt_integration.hpp"
 #include "vast/detail/operators.hpp"
 #include "vast/detail/overload.hpp"
 #include "vast/offset.hpp"
@@ -29,8 +28,6 @@
 #include <caf/none.hpp>
 #include <caf/optional.hpp>
 #include <caf/variant.hpp>
-
-#include <fmt/format.h>
 
 #include <chrono>
 #include <filesystem>
@@ -398,262 +395,3 @@ struct hash<vast::data> {
 };
 
 } // namespace std
-
-namespace fmt {
-
-template <>
-struct formatter<vast::map::value_type>
-  : public vast::detail::vast_formatter_base {
-  template <typename FormatContext>
-  auto format(const vast::map::value_type& v, FormatContext& ctx) {
-    if (presentation == 'a')
-      return fmt::format_to(ctx.out(), "{:a} -> {:a}", v.first, v.second);
-    else
-      return fmt::format_to(ctx.out(), "{:j}: {:j}", v.first, v.second);
-  }
-};
-
-template <>
-struct formatter<vast::record::value_type>
-  : public vast::detail::vast_formatter_base {
-  template <typename FormatContext>
-  auto format(const vast::record::value_type& v, FormatContext& ctx) {
-    if (presentation == 'a')
-      return fmt::format_to(ctx.out(), "{}: {:a}", v.first, v.second);
-    else
-      return fmt::format_to(ctx.out(), "{}: {:j}", v.first, v.second);
-  }
-};
-
-/// Definition of fmt-formatting rules for vast::data.
-template <>
-struct formatter<vast::data> : public vast::detail::vast_formatter_base {
-  using ascii_escape_string
-    = vast::detail::escaped_string_view<vast::detail::print_escaper_functor>;
-  using json_escape_string
-    = vast::detail::escaped_string_view<vast::detail::json_escaper_functor>;
-
-  template <typename Output>
-  struct ascii_visitor {
-    Output out_;
-
-    ascii_visitor(Output out) : out_{std::move(out)} {
-    }
-
-    auto operator()(caf::none_t) {
-      return format_to(out_, "nil");
-    }
-    auto operator()(bool b) {
-      return format_to(out_, b ? "T" : "F");
-    }
-    auto operator()(vast::duration d) {
-      return format_to(out_, "{}",
-                       vast::detail::fmt_wrapped<vast::duration>{d});
-    }
-    auto operator()(const vast::time& t) {
-      return format_to(out_, "{}",
-                       vast::detail::fmt_wrapped<vast::time>{t});
-    }
-    auto operator()(const std::string& s) {
-      return (*this)(ascii_escape_string{s});
-    }
-    auto operator()(const vast::list& xs) {
-      return format_to(out_, "[{:a}]", join(xs, ", "));
-    }
-    auto operator()(const vast::map& xs) {
-      return format_to(out_, "{{{:a}}}", join(xs, ", "));
-    }
-    auto operator()(const vast::record& xs) {
-      return format_to(out_, "<{:a}>", join(xs, ", "));
-    }
-
-    template <class T>
-    auto operator()(const T& x) {
-      return format_to(out_, "{}", x);
-    }
-  };
-
-  template <typename Output>
-  struct json_visitor_base {
-    Output out_;
-    json_visitor_base(Output out) : out_{std::move(out)} {}
-
-    auto operator()(caf::none_t) {
-      return format_to(out_, "null");
-    }
-    auto operator()(bool b) {
-      return format_to(out_, b ? "true" : "false");
-    }
-    auto operator()(vast::duration d) {
-      return format_to(out_, "\"{}\"",
-                       vast::detail::fmt_wrapped<vast::duration>{d});
-    }
-    auto operator()(const vast::time& t) {
-      return format_to(out_, "\"{}\"",
-                       vast::detail::fmt_wrapped<vast::time>{t});
-    }
-    auto operator()(const std::string& s) {
-      return (*this)(json_escape_string{s});
-    }
-
-    template <class T>
-    auto operator()(const T& x) {
-      return format_to(out_, "{}", x);
-    }
-  };
-
-  template <typename Output, class PrintTraits>
-  struct json_visitor : json_visitor_base<Output>, PrintTraits {
-    using super = json_visitor_base<Output>;
-    PrintTraits print_traits_;
-    json_visitor(Output out, PrintTraits print_traits)
-      : super{out}, print_traits_{std::move(print_traits)} {
-    }
-
-    template <class T>
-    auto operator()(const T& x) {
-      return static_cast<super&>(*this)(x);
-    }
-
-    auto operator()(const vast::list& xs) {
-      *this->out_++ = '[';
-      if (!xs.empty()) {
-        print_traits_.inc_indent();
-        bool is_first = true;
-        for(const auto & x: xs) {
-          if (is_first)
-            is_first = false;
-          else
-            *this->out_++ = ',';
-          print_traits_.format_indent(this->out_);
-          this->out_ = caf::visit(*this, x);
-        }
-        print_traits_.dec_indent();
-        print_traits_.format_indent(this->out_);
-      }
-      *this->out_ = ']';
-      return this->out_;
-    }
-
-    auto operator()(const vast::map& xs) {
-      *this->out_++ = '[';
-      if (!xs.empty()) {
-        print_traits_.inc_indent();
-        bool is_first = true;
-        for(const auto & x: xs) {
-          if (is_first)
-            is_first = false;
-          else
-            *this->out_++ = ',';
-
-          print_traits_.format_indent(this->out_);
-          *this->out_++ = '{';
-          print_traits_.inc_indent();
-          print_traits_.format_indent(this->out_);
-
-          print_traits_.format_field_start(this->out_, "key");
-          this->out_ = caf::visit(*this, x.first);
-          *this->out_++ = ',';
-          print_traits_.format_indent(this->out_);
-          print_traits_.format_field_start(this->out_, "value");
-          this->out_ = caf::visit(*this, x.second);
-
-          print_traits_.dec_indent();
-          print_traits_.format_indent(this->out_);
-          *this->out_++ = '}';
-        }
-        print_traits_.dec_indent();
-        print_traits_.format_indent(this->out_);
-      }
-      *this->out_ = ']';
-      return this->out_;
-    }
-
-    auto operator()(const vast::record& xs) {
-      *this->out_++ = '{';
-      if (!xs.empty()) {
-        print_traits_.inc_indent();
-        bool is_first = true;
-        for(const auto & x: xs) {
-          if (is_first)
-            is_first = false;
-          else
-            *this->out_++ = ',';
-          print_traits_.format_indent(this->out_);
-          print_traits_.format_field_start(this->out_, x.first);
-          this->out_ = caf::visit(*this, x.second);
-        }
-        print_traits_.dec_indent();
-        print_traits_.format_indent(this->out_);
-      }
-      *this->out_ = '}';
-      return this->out_;
-    }
-  };
-
-  template <bool RemoveSpaces>
-  struct ndjson_print_traits {
-    template <class... Ts>
-    constexpr void dec_indent(Ts&&...) const noexcept {
-    }
-    template <class... Ts>
-    constexpr void inc_indent(Ts&&...) const noexcept {
-    }
-    template <class Output>
-    constexpr void format_indent(Output& out) const noexcept {
-      if constexpr (!RemoveSpaces) {
-        *out++ = ' ';
-      }
-    }
-    template <class Output>
-    void format_field_start(Output& out, std::string_view name) {
-      out = format_to(out, RemoveSpaces? "{}:" : "{}: ", json_escape_string{name});
-    }
-  };
-
-  struct json_print_traits {
-    int indent_size;
-    int current_indent{0};
-
-    template <class... Ts>
-    constexpr void dec_indent(Ts&&...) noexcept {
-      --current_indent;
-    }
-    template <class... Ts>
-    constexpr void inc_indent(Ts&&...) noexcept {
-      ++current_indent;
-    }
-    template <class Output>
-    constexpr void format_indent(Output& out) const noexcept {
-      out = format_to(out, "\n{:<{}}", "", current_indent*indent_size);
-
-    }
-    template <class Output>
-    void format_field_start(Output& out, std::string_view name) {
-      out = format_to(out, "{}: ", json_escape_string{name});
-    }
-  };
-
-  template <typename FormatContext>
-  auto format(const vast::data& x, FormatContext& ctx) {
-    if (presentation == 'a') {
-      ascii_visitor f{ctx.out()};
-      return caf::visit(f, x);
-    } else if (presentation == 'j') {
-      auto do_format = [&x](auto f){ return caf::visit(f, x); };
-      if (ndjson) {
-        if (remove_spaces)
-          return do_format(json_visitor{ctx.out(), ndjson_print_traits<true>{}});
-        else
-          return do_format(json_visitor{ctx.out(), ndjson_print_traits<false>{}});
-      } else
-        return do_format(json_visitor{ctx.out(), json_print_traits{indent}});
-    } else if (presentation == 'j') {
-
-    }
-
-    return ctx.out();
-  }
-};
-
-} // namespace fmt
